feat: minimap and loc handling

+++ src/js/jagex2/config/IdkType.ts
@@ -19,10 +19,10 @@ export default class IdkType extends ConfigType {
     // ----
 
     type: number = -1;
-    models: Uint16Array | null = null;
-    heads: Uint16Array = new Uint16Array(5).fill(-1);
-    recol_s: Uint16Array = new Uint16Array(6);
-    recol_d: Uint16Array = new Uint16Array(6);
+    models: Int32Array | null = null;
+    heads: Int32Array = new Int32Array(5).fill(-1);
+    recol_s: Int32Array = new Int32Array(6);
+    recol_d: Int32Array = new Int32Array(6);
     disable: boolean = false;
 
     decode = (_index: number, code: number, dat: Packet): void => {
@@ -30,7 +30,7 @@ export default class IdkType extends ConfigType {
             this.type = dat.g1;
         } else if (code === 2) {
             const count: number = dat.g1;
-            this.models = new Uint16Array(count);
+            this.models = new Int32Array(count);
 
             for (let i: number = 0; i < count; i++) {
                 this.models[i] = dat.g2;
